// src/main/resources/APachas/src/app/services/entities/User.ts
export class User {
    userId: number;
    userName: string;
    userSurname: string;
    userLogin: string;
    userPassword: string;
    userEmail: string;
    userBirthday: Date;
    userPhoto: string;
    roles: string;
    permissions: string;
    userActive: boolean;
    userCreation: Date;
    userRemoval: Date;
    userNotify: boolean;
}